Removed the explore section
After making some of the changes described in Issue #64 it
became clear that the current design for the explore section
is not suited for the newly discovered requirements.

The list of states is too long; there are many candidates
with ads that were never aired in our markets, resulting
in long, blank lists.

Furthermore, the air counts are not broken up in terms
of race, meaning candidates currently running for senate
who originally ran for president will have their presidential
ads counted in the senate race graphs.