Removed relative font sizes
Relative font size is fine, but it was inconsistantly applied in this
theme and also used for things like margin and padding.  Ultimately
this resulted in an unpredicatble user experience.  use of em and rem
has been removed from the theme; it can be added back if needed,
but only if we approach it for everything and make sure we document
the new expectations for the code base.